requirements.txt
django==4.1.1
django-cors-headers==3.13.0
Pillow==9.3.0
django-summernote==0.8.20.0
django-extensions==3.2.1
pydotplus==2.0.2
gunicorn==20.1.0
djangorestframework==3.14.0